backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
python-multipart>=0.0.5
openai>=1.0.0
numpy>=1.21.0
python-dotenv>=0.19.0